Update #1 Client/Server Networking
A commond Shared "GameSocket" is used by both Clients and Servers to communicate to each other. It contains a running thread to collected recieved packets as well as a nice and friendly send wrapper. Ideally a Client has a wrapper for a GameSocket for their own needs, and the Server contains some sort of map to keep track of all its users collective Sockets. I believe I've created the basic implementation for this.

Future Updates will include wiring the current System to work ( Which it should ). Wire the current System to some sort of log. Wire the current System ( Server Side ) to be able to maintain some sort of Player Object.