Update #5 Shaders!? Backup
So my current quest since I finally finished the Network backbone, is to get "Visual Effects" using Shaders working. While Slick2d ( The game Engine I use ) allows us to display sprites, backgrounds, and text just fine, there isn't any good support for Shaders using OpenGL that I can find. So, I've decided to backup our project, and work on implementing LWJGL ( The library Slick is made on ) to work with our shaders. It might be possible we discard Slick2d and rewrite the game with LWJGL. We are early enough along for it to be a worthwhile change