Update #39 The Culling
The FrustrumCullingFilter class offers the ability to mark both GameItems, and BLOCKS, out of the field of view of a Camera. Using this, remarkable improvement can be made towards rendering optimization.

The next update will likely focus on working on not rendering block sides that aren't visible using the camera's rotations