Update #55 Rewrite ChunkIO, Entities, Blocks
All the old classes have been moved to a folder where they can be crossed reference for the future while re-organizing this library.

ChunkIO uses the new ChunkFormat design, which uses Huffman tree's and a Pallete to save all the Region to file. This presents a few issues, such as the removal of the "NULL" to designate Air blocks. and use actual blocks. Chunks should follow a 256x256 sizing, which will allow for ideal amounts of repeating, and also allow the maximum use of bytes in the least amount of files possible. The usage of "Modfifiers" for deconstructing files has not been ruled out either, and may be implemented further in future iterations. Ideally, this format size will in the long term be better spacewise, and will be tested against the old format ( Air bytes representing future airs ) to assure its long term usage.